+++ packages/replay/test/integration/beforeAddRecordingEvent.test.ts
@@ -5,6 +5,7 @@ import * as SentryUtils from '@sentry/utils';
 import type { Replay } from '../../src';
 import type { ReplayContainer } from '../../src/replay';
 import { clearSession } from '../../src/session/clearSession';
+import type { EventType } from '../../src/types';
 import * as SendReplayRequest from '../../src/util/sendReplayRequest';
 import { BASE_TIMESTAMP, mockRrweb, mockSdk } from '../index';
 import { useFakeTimers } from '../utils/use-fake-timers';
@@ -37,7 +38,7 @@ describe('Integration | beforeAddRecordingEvent', () => {
     ({ replay, integration } = await mockSdk({
       replayOptions: {
         beforeAddRecordingEvent: event => {
-          const eventData = event.data as Record<string, any>;
+          const eventData = event.data;
 
           if (eventData.tag === 'breadcrumb' && eventData.payload.category === 'ui.click') {
             return {
@@ -53,8 +54,8 @@ describe('Integration | beforeAddRecordingEvent', () => {
           }
 
           // This should not do anything because callback should not be called
-          // for `event.type != 5`
-          if (event.type === 2) {
+          // for `event.type != 5` - but we guard anyhow to be safe
+          if ((event.type as EventType) === 2) {
             return null;
           }